+++ axelor-base/src/main/java/com/axelor/apps/base/exceptions/IExceptionMessage.java
@@ -377,4 +377,7 @@ public interface IExceptionMessage {
   public static final String ADVANCED_IMPORT_LOG_8 = /*$$(*/
       "Missing data for required fields" /*)*/;
   public static final String ADVANCED_IMPORT_LOG_9 = /*$$(*/ "Invalid type of data" /*)*/;
+  public static final String ADVANCED_IMPORT_RESET = /*$$(*/
+      "Reset imported data successfully" /*)*/;
+  public static final String ADVANCED_IMPORT_NO_RESET = /*$$(*/ "No imported data to reset" /*)*/;
 }