+++ axelor-base/src/main/java/com/axelor/apps/base/service/advanced/imports/DataImportService.java
@@ -18,12 +18,16 @@
 package com.axelor.apps.base.service.advanced.imports;
 
 import com.axelor.apps.base.db.AdvancedImport;
+import com.axelor.apps.base.db.FileTab;
 import com.axelor.exception.AxelorException;
 import com.axelor.meta.db.MetaFile;
 import java.io.IOException;
+import java.util.Map;
 
 public interface DataImportService {
 
   public MetaFile importData(AdvancedImport advanceImport)
       throws IOException, AxelorException, ClassNotFoundException;
+
+  public Map<String, Object> createJsonContext(FileTab fileTab);
 }